moved profile footer to Character component folder
