+++ components/Character/ProfileFooter/index.tsx
